chore: code consolidation
