--- crates/intrinsic-test/src/common/supporting_test.rs
@@ -1,10 +1,0 @@
-use crate::common::cli::ProcessedCli;
-
-/// Architectures must support this trait
-/// to be successfully tested.
-pub trait SupportedArchitectureTest {
-    fn create(cli_options: ProcessedCli) -> Self;
-    fn build_c_file(&self) -> bool;
-    fn build_rust_file(&self) -> bool;
-    fn compare_outputs(&self) -> bool;
-}
